pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etl-toolkit"
version = "1.0.0"
description = "Lightweight, Observable ETL Framework for Python"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Kai Song", email = "kaisong.uk@outlook.com"}]
keywords = ["etl", "data-pipeline", "data-quality", "python"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = ["psycopg2-binary>=2.9.9"]

[project.optional-dependencies]
dev = ["pytest>=7.4", "black>=23.0", "mypy>=1.5", "ruff>=0.1"]
mysql = ["pymysql>=1.1.0"]
pandas = ["pandas>=2.0"]
all = ["etl-toolkit[dev,mysql,pandas]"]

[project.urls]
Homepage = "https://github.com/KaiSong-UK/etl-toolkit"

[tool.setuptools.packages.find]
where = ["src"]
